orchestrator/requirements_lock.txt
acme==2.9.0
ansible==9.2.0
ansible-core==2.16.3
apache-libcloud==3.4.1
argcomplete==3.1.4
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
blinker==1.7.0
borgbackup==1.2.8
Brlapi==0.8.5
certbot==2.9.0
certbot-apache==2.9.0
certbot-nginx==2.9.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==24.4.1
cockpit @ file:///build/cockpit-mHshv0/cockpit-314/tmp/wheel/cockpit-314-py3-none-any.whl#sha256=5ee1902356c6db04f12a0b33e9cef9bf9201ffc4523b40d6c79c9b2e54e06337
colorama==0.4.6
command-not-found==0.3
ConfigArgParse==1.7
configobj==5.0.8
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
dnspython==2.6.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
fail2ban==1.0.2
gyp==0.1
httplib2==0.20.4
idna==3.6
iniconfig==1.1.1
j2cli==0.3.12b0
Jinja2==3.1.2
jmespath==1.0.1
josepy==1.14.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.29.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.0.3
netaddr==0.8.0
netifaces==0.11.0
ntlm-auth==1.5.0
oauthlib==3.2.2
olefile==0.46
packaging==24.0
paramiko==2.12.0
parsedatetime==2.6
passlib==1.7.4
pathspec==0.12.1
pcp==5.0
pexpect==4.9.0
pillow==10.2.0
pipx==1.4.3
platformdirs==4.2.0
pluggy==1.4.0
prometheus-client==0.19.0
psutil==5.9.8
ptyprocess==0.7.0
pyasyncore==1.0.2
pycairo==1.25.1
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.48.2
PyICU==2.12
pyinotify==0.9.6
PyJWT==2.7.0
pykerberos==1.1.14
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyRFC3339==1.1
pyrsistent==0.20.0
pyserial==3.5
pytest==7.4.4
python-apt==2.7.7+ubuntu4
python-augeas==0.5.0
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.1
pytz==2024.1
pywinrm==0.4.3
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
requests-ntlm==1.1.0
resolvelib==1.0.1
rich==13.7.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
selinux @ file:///build/libselinux-IVxPh3/libselinux-3.5/src
setuptools==68.1.2
simplejson==3.19.2
six==1.16.0
ssh-import-id==5.11
systemd-python==235
texttable==1.6.7
toml==0.10.2
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
userpath==1.9.1
wadllib==1.3.6
websocket-client==1.7.0
wheel==0.42.0
xdg==5
xkit==0.0.0
xmltodict==0.13.0
yamllint==1.33.0
yq==3.1.0
